refactor: port travel-mode to svelte
